TotalDifficulty is a BigInteger
Don't try and cast total difficulty down to a long because it will
overflow long in a reasonable timeframe (mainnet already has).

+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/AdminNodeInfo.java
@@ -118,7 +118,7 @@ public class AdminNodeInfo implements JsonRpcMethod {
                 "config",
                 genesisConfigOptions.asMap(),
                 "difficulty",
-                chainHead.getTotalDifficulty().toLong(),
+                chainHead.getTotalDifficulty().toBigInteger(),
                 "genesis",
                 blockchainQueries.getBlockHashByNumber(0).get().toString(),
                 "head",
